[hotfix][checkpoint] Extract CheckpointPlan to be an interface
